206: move more functions to graphUtils
