+++ src/navmesh/graph.js
@@ -1,31 +1,6 @@
 import _ from 'lodash';
 import { assert } from '../utils/asserts';
-import { findUpperAndLowerPoints, threePointsInLine } from './graphUtils';
-
-export function determinant(matrix) {
-  const N = matrix.length;
-  for (let i = 0; i < N; i += 1) {
-    assert(matrix[i].length === N, 'input matrix should be NxN');
-  }
-  let sum = 0;
-  // Recursive base-case, a single element
-  if (N === 1) return matrix[0][0];
-  for (let j = 0; j < N; j += 1) {
-    // Create a sub-matrix which do not contain elements in the 0th row or the jth column
-    const subMatrix = [];
-    for (let k = 1; k < N; k += 1) {
-      const row = [];
-      for (let l = 0; l < N; l += 1) {
-        if (l !== j) row.push(matrix[k][l]);
-      }
-      subMatrix.push(row);
-    }
-    // Alternate between +/- the determinant of the sub-matrix
-    sum += ((j % 2) ? -1 : 1) * matrix[0][j] * determinant(subMatrix);
-  }
-  return sum;
-}
-
+import { findUpperAndLowerPoints, threePointsInLine, detD, detH } from './graphUtils';
 
 /**
  * Represents an x, y pixel location on the tagpro map. Used as vertices to define polygons.
@@ -74,20 +49,6 @@ export class Point {
   toString() {
     return `x: ${this.x}, y: ${this.y}`;
   }
-}
-
-/**
- * @param {Point} A
- * @param {Point} B
- * @param {Point} P
- * @returns {number} a number which is positive iff P is on the left of the edge AB
- */
-function detD(A, B, P) {
-  return determinant([
-    [A.x, A.y, 1],
-    [B.x, B.y, 1],
-    [P.x, P.y, 1],
-  ]);
 }
 
 /**
@@ -176,14 +137,8 @@ export function sortCounterClockwise(points) {
  */
 export function isLegal(insertedPoint, e, oppositePoint) {
   const [A, B, C] = sortCounterClockwise([insertedPoint, e.p1, e.p2]);
-  const D = oppositePoint;
-  const matrix = [
-    [A.x, A.y, (A.x ** 2) + (A.y ** 2), 1],
-    [B.x, B.y, (B.x ** 2) + (B.y ** 2), 1],
-    [C.x, C.y, (C.x ** 2) + (C.y ** 2), 1],
-    [D.x, D.y, (D.x ** 2) + (D.y ** 2), 1],
-  ];
-  return determinant(matrix) <= 0;
+  const E = oppositePoint;
+  return detH(A, B, C, E) <= 0;
 }
 
 
@@ -518,6 +473,7 @@ export class TGraph extends Graph {
   }
 
   addFixedEdge(e) {
+    assert(this.hasVertex(e.p1) && this.hasVertex(e.p2));
     super.addEdge(e.p1, e.p2);
     this.fixedAdj[e.p1].push(e.p2);
     this.fixedAdj[e.p2].push(e.p1);
