+++ src/navmesh/graphUtils.js
@@ -1,5 +1,7 @@
 import _ from 'lodash';
-import { Point, pointsOnSameSide } from '../navmesh/graph';
+
+import { Point, pointsOnSameSide } from './graph';
+import { assert } from '../utils/asserts';
 
 const CLEARANCE = 19;
 
@@ -61,6 +63,7 @@ export function findUpperAndLowerPoints(intersectingTriangles, e) {
  * @returns {boolean} true if all points are colinear
  */
 export function threePointsInLine(p1, p2, p3) {
+  if (p1.equal(p2) || p1.equal(p3) || p2.equal(p3)) return true;
   const x1 = p2.x - p1.x;
   const x2 = p2.x - p3.x;
   const y1 = p2.y - p1.y;
@@ -101,4 +104,61 @@ export function getClearancePoint(cornerPoint, prevPoint, nextPoint) {
 
   // Insert other point to path
   return cornerPoint.add(normal.times(CLEARANCE));
+}
+
+
+export function determinant(matrix) {
+  const N = matrix.length;
+  for (let i = 0; i < N; i += 1) {
+    assert(matrix[i].length === N, 'input matrix should be NxN');
+  }
+  let sum = 0;
+  // Recursive base-case, a single element
+  if (N === 1) return matrix[0][0];
+  for (let j = 0; j < N; j += 1) {
+    // Create a sub-matrix which do not contain elements in the 0th row or the jth column
+    const subMatrix = [];
+    for (let k = 1; k < N; k += 1) {
+      const row = [];
+      for (let l = 0; l < N; l += 1) {
+        if (l !== j) row.push(matrix[k][l]);
+      }
+      subMatrix.push(row);
+    }
+    // Alternate between +/- the determinant of the sub-matrix
+    sum += ((j % 2) ? -1 : 1) * matrix[0][j] * determinant(subMatrix);
+  }
+  return sum;
+}
+
+
+/**
+ * @param {Point} A
+ * @param {Point} B
+ * @param {Point} P
+ * @returns {number} a number which is positive iff P is on the left of the edge AB
+ */
+export function detD(A, B, P) {
+  return determinant([
+    [A.x, A.y, 1],
+    [B.x, B.y, 1],
+    [P.x, P.y, 1],
+  ]);
+}
+
+
+/**
+ * @param {Point} A
+ * @param {Point} B
+ * @param {Point} C
+ * @param {Point} E
+ * @returns {number} a number which is positive iff D is inside the circumcircle of points A, B, C
+ */
+export function detH(A, B, C, D) {
+  return determinant([
+    [A.x, A.y, (A.x ** 2) + (A.y ** 2), 1],
+    [B.x, B.y, (B.x ** 2) + (B.y ** 2), 1],
+    [C.x, C.y, (C.x ** 2) + (C.y ** 2), 1],
+    [D.x, D.y, (D.x ** 2) + (D.y ** 2), 1],
+  ]);
 }
