+++ test/navmesh/graph.spec.js
@@ -1,6 +1,5 @@
 import test from 'tape';
 import {
-  determinant,
   isLegal,
   sortCounterClockwise,
   pointsOnSameSide,
@@ -9,6 +8,7 @@ import {
   Triangle,
   TGraph,
 } from '../../src/navmesh/graph';
+import { determinant } from '../../src/navmesh/graphUtils';
 
 
 test('determinant returns correct value for a 3x3', t => {
